backend/requirements.txt
# Slack Integration
slack-sdk==3.27.1
slack-bolt==1.18.1

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25  # Optional: ORM if you prefer

# Environment & Configuration
python-dotenv==1.0.1

# Async Support
aiohttp==3.9.3
asyncio==3.4.3

# Utilities
requests==2.31.0
urllib3==2.2.0

# Logging & Monitoring
python-json-logger==2.0.7

# CLI
click==8.1.7  # For building CLI scripts

# AI
chromadb==0.4.22
anthropic>=0.18.0
numpy<2.0.0  # ChromaDB compatibility

# Task Scheduling
apscheduler==3.11.1

# FastAPI & Web
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
cryptography>=3.4.8  # For Fernet encryption

# File handling
python-magic==0.4.27
PyPDF2==3.0.1
python-docx==1.1.0
aiofiles==23.2.1

# Development (optional)
pytest==8.0.0
pytest-asyncio==0.23.5
black==24.2.0
flake8==7.0.0
